Release SOD 0.70
Multi-species/multi-nary substitution, molecules (@NAME) and vacancies
(%NAME), entropy statistics, dual-level post-processing scripts,
thermodynamics units in statsod/gcstatsod, Fortran 2003/2008 quality
improvements (real64 throughout, character(len=N), stop 1, i0 format),
new example06 (MgH2 multi-species VASP), updated README and VERSIONS.

+++ src/cell.f90
@@ -19,44 +19,44 @@
 !******************************************************************************
 
 subroutine cell(cellvector, a, b, c, alpha, beta, gamma)
+  use iso_fortran_env, only: real64
   implicit none
 
-  real, parameter:: pi = acos(-1.0), degtorad = pi/180.0
+  real(real64), parameter :: pi = acos(-1.0_real64), degtorad = pi/180.0_real64
 
-  real, dimension(3, 3):: cellvector
-  real :: a, b, c, alpha, beta, gamma, alp, bet, gam, cosa, cosb, cosg, sing
-  real :: trm1
+  real(real64), dimension(3, 3) :: cellvector
+  real(real64) :: a, b, c, alpha, beta, gamma, alp, bet, gam, cosa, cosb, cosg, sing
+  real(real64) :: trm1
 
-  if (alpha == 90.0) then
-    cosa = 0.0
+  if (alpha == 90.0_real64) then
+    cosa = 0.0_real64
   else
     alp = alpha*degtorad
     cosa = cos(alp)
   end if
-  if (beta == 90.0) then
-    cosb = 0.0
+  if (beta == 90.0_real64) then
+    cosb = 0.0_real64
   else
     bet = beta*degtorad
     cosb = cos(bet)
   end if
-  if (gamma == 90.0) then
-    sing = 1.0
-    cosg = 0.0
+  if (gamma == 90.0_real64) then
+    sing = 1.0_real64
+    cosg = 0.0_real64
   else
     gam = gamma*degtorad
     sing = sin(gam)
     cosg = cos(gam)
   end if
-  cellvector(2, 1) = 0.0
-  cellvector(3, 1) = 0.0
-  cellvector(3, 2) = 0.0
+  cellvector(2, 1) = 0.0_real64
+  cellvector(3, 1) = 0.0_real64
+  cellvector(3, 2) = 0.0_real64
   cellvector(1, 1) = a
   cellvector(1, 2) = b*cosg
   cellvector(2, 2) = b*sing
   cellvector(1, 3) = c*cosb
   cellvector(2, 3) = c*(cosa - cosg*cosb)/sing
   trm1 = cellvector(2, 3)/c
-  cellvector(3, 3) = c*sqrt(1.0 - cosb**2 - trm1**2)
+  cellvector(3, 3) = c*sqrt(1.0_real64 - cosb**2 - trm1**2)
   return
 end
-
