+++ src/factorials.f90
@@ -1,15 +1,17 @@
 function pbinomial(nsubs, npos, x)
+  use iso_fortran_env, only: int64, real64
   implicit none
   integer :: nsubs, npos
-  integer(kind=8) :: combinations
-  real*8 :: x, pbinomial
+  integer(int64) :: combinations
+  real(real64) :: x, pbinomial
   pbinomial = combinations(nsubs, npos)*(x**nsubs)*((1 - x)**(npos - nsubs))
 end function pbinomial
 
 function combinations(nsubs, npos)
+  use iso_fortran_env, only: int64, real64
   implicit none
   integer :: nsubs, npos, k, p
-  integer(kind=8)::  factorial, combinations, ratio
+  integer(int64)::  factorial, combinations, ratio
 
   if (nsubs <= npos/2) then
     k = nsubs
@@ -24,9 +26,10 @@ function combinations(nsubs, npos)
 end function combinations
 
 recursive function factorial(n) result(aux)
+  use iso_fortran_env, only: int64, real64
   implicit none
   integer, intent(in) :: n
-  integer(kind=8):: aux
+  integer(int64):: aux
   if (n == 0) then
     aux = 1
   else
