+++ src/momenta.f90
@@ -1,7 +1,8 @@
 function momentaa0(nsubsmax, npos, x)
+  use iso_fortran_env, only: real64
   implicit none
   integer :: nsubs, nsubsmax, npos
-  real*8 :: x, pbinomial, momentaa0
+  real(real64) :: x, pbinomial, momentaa0
   momentaa0 = 0.0
   do nsubs = nsubsmax + 1, npos
     momentaa0 = momentaa0 + pbinomial(nsubs, npos, x)
@@ -9,9 +10,10 @@ function momentaa0(nsubsmax, npos, x)
 end function momentaa0
 
 function momentaa1(nsubsmax, npos, x)
+  use iso_fortran_env, only: real64
   implicit none
   integer :: nsubs, nsubsmax, npos
-  real*8 :: x, pbinomial, momentaa1
+  real(real64) :: x, pbinomial, momentaa1
   momentaa1 = 0.0
   do nsubs = nsubsmax + 1, npos
     momentaa1 = momentaa1 + pbinomial(nsubs, npos, x)*nsubs
@@ -19,9 +21,10 @@ function momentaa1(nsubsmax, npos, x)
 end function momentaa1
 
 function momentaa2(nsubsmax, npos, x)
+  use iso_fortran_env, only: real64
   implicit none
   integer :: nsubs, nsubsmax, npos
-  real*8 :: x, pbinomial, momentaa2
+  real(real64) :: x, pbinomial, momentaa2
   momentaa2 = 0.0
   do nsubs = nsubsmax + 1, npos
     momentaa2 = momentaa2 + pbinomial(nsubs, npos, x)*(nsubs**2)
@@ -29,9 +32,10 @@ function momentaa2(nsubsmax, npos, x)
 end function momentaa2
 
 function momentab0(nsubsmin, npos, x)
+  use iso_fortran_env, only: real64
   implicit none
   integer :: nsubs, nsubsmin, npos
-  real*8 :: x, pbinomial, momentab0
+  real(real64) :: x, pbinomial, momentab0
   momentab0 = 0.0
   do nsubs = 0, nsubsmin - 1
     momentab0 = momentab0 + pbinomial(nsubs, npos, x)
@@ -39,9 +43,10 @@ function momentab0(nsubsmin, npos, x)
 end function momentab0
 
 function momentab1(nsubsmin, npos, x)
+  use iso_fortran_env, only: real64
   implicit none
   integer :: nsubs, nsubsmin, npos
-  real*8 :: x, pbinomial, momentab1
+  real(real64) :: x, pbinomial, momentab1
   momentab1 = 0.0
   do nsubs = 0, nsubsmin - 1
     momentab1 = momentab1 + pbinomial(nsubs, npos, x)*nsubs
@@ -49,9 +54,10 @@ function momentab1(nsubsmin, npos, x)
 end function momentab1
 
 function momentab2(nsubsmin, npos, x)
+  use iso_fortran_env, only: real64
   implicit none
   integer :: nsubs, nsubsmin, npos
-  real*8 :: x, pbinomial, momentab2
+  real(real64) :: x, pbinomial, momentab2
   momentab2 = 0.0
   do nsubs = 0, nsubsmin - 1
     momentab2 = momentab2 + pbinomial(nsubs, npos, x)*(nsubs**2)
